gbsim/loopback_test: add unicast, multicast bitmap support
Feature add which enables the ability to select a bit-mask of connections
to run when executing a loopback test set. Here we add support to the
user-space binary to allow specification of the connection inclusion
bitmask on the utility command line.

-m bitmask is added where bitmask is classic binary counting
   i.e. 1 for 1st connection 5 for connections 1 and 3, 8 for
   connection 4 in isolation or 12 for connections 3 and 4

This is a feature request from the firmware people which ought to enable
masking of connections and therefore should negate a requirement to
recompile firmware in order to add or remove connections from a particular
test set.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>