gbsim/loopback_test: align names to sysfs and debugfs datapoints
sysfs and debugfs entries have been updated to match more clearly with the
greybus namespace. This includes putting data that pertains to all cports
in a particular module under names that at a 'module level' or at a
'connection level' e.g.

/sys/bus/greybus/devices/endo0:3/error_dev
/sys/bus/greybus/devices/endo0:3:3:3:4/error_con

In order to gather this data the user-space utility needs an update to
facilitate.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>