gbsim: add lights support
Add lights support to gbsim to test greybus implementation of the
lights class protocol. This code will create 3 lights devices with 4
channels each. The first 2 are simple LED devices and the last one a
mix of LED and flash type channels.

Setting the brightness on any channel of a light to 254 will trigger a
test hack to send a reconfigure event. The greybus kernel side should
release the all channels in that light and start a new configuration.
This can be validate by checking the timestamps of the sysfs entries
of the channels.

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>