gbsim/loopback_test: make module debugfs data-point non-module dependent
Having a module identifier as part of the path of debug/sysfs entries
pertaining to the entire set of data coming from loopback makes no sense.
Aside from not being logical - it also drives the kernel to populate
unnecessary data points in /sys on a per-module basis. For user-space not
caring about the module-id in the path means less work. So for both kernel
and user space this is a win.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>